[ALPS05376395] wlan: enlarge wifi cfg size
Enlarge wifi cfg size to 400 due to already full.
when size is full, it will print error log cause log too much.

MTK-Commit-Id: 28091d0ccd13edfa2cee7268b63954bc944ed754

CR-Id: ALPS05376395
Feature: [Module]Wi-Fi Driver
(cherry picked from commit f015970dd582be3f8b79e5825e178f0aa1290e59)

+++ common/wlan_lib.c
@@ -12294,7 +12294,7 @@ void
 wlanResoreEmCfgSetting(IN struct ADAPTER *
 	prAdapter)
 {
-	uint8_t i;
+	uint32_t i;
 
 	for (i = 0; i < WLAN_CFG_ENTRY_NUM_MAX; i++) {
 
@@ -12319,7 +12319,7 @@ void
 wlanBackupEmCfgSetting(IN struct ADAPTER *
 	prAdapter)
 {
-	uint8_t i;
+	uint32_t i;
 	struct WLAN_CFG_ENTRY *prWlanCfgEntry = NULL;
 
 	kalMemZero(&g_rEmCfgBk, sizeof(g_rEmCfgBk));
@@ -12358,7 +12358,7 @@ void
 wlanCleanAllEmCfgSetting(IN struct ADAPTER *
 	prAdapter)
 {
-	uint8_t i;
+	uint32_t i;
 	struct WLAN_CFG_ENTRY *prWlanCfgEntry = NULL;
 
 	for (i = 0; i < WLAN_CFG_ENTRY_NUM_MAX; i++) {
